Добавлены фильтры (#90)
* add filter-sort

* add sort and filter for category

* fix app

* fix status and add filters for search

+++ db/migrations/000001__create_tables.up.sql
@@ -131,7 +131,7 @@ CREATE TABLE IF NOT EXISTS bazaar.product
     price             NUMERIC(12, 2) CHECK (price >= 0) NOT NULL,
     quantity          INT CHECK (quantity >= 0)         NOT NULL,
     updated_at        TIMESTAMPTZ                        DEFAULT now(),
-    rating            INT CHECK (rating BETWEEN 0 AND 5) DEFAULT 0,
+    rating            FLOAT CHECK (rating BETWEEN 0 AND 5) DEFAULT 0,
     reviews_count     INT CHECK (reviews_count >= 0)     DEFAULT 0 --trigger
 );
